refactor: corrected props usage and file placement
